protect agsint indexing error

+++ src/supremacy/vehicles.py
@@ -224,6 +224,9 @@ class Ship(Vehicle):
             print("No land found around ship, cannot build base on water!")
             return
         yy, xx = np.where(local_views[0] == 1)
+        if not np.shape(xx):
+            print("Error findind land around ship....")
+            return
         uid = player.build_base(x=x + xx[0] - 1, y=y + yy[0] - 1)
         player.transformed_ships.append(self.uid)
         self.avatar.delete()
